refactor: reorganize imports and enhance logging functionality
- Moved `startServer` to a new `extensions` directory and updated its import path.
- Refactored logger-related imports to use a new `logger` module.
- Introduced new helper functions for logging, including color mapping and duration formatting.
- Added new helper functions for constructing log messages and handling HTTP errors.
- Implemented file and console transport for logging output.
- Created interfaces for better type definitions and improved code organization.
