+++ src/helpers/duration.ts
